테스트 코드 작성(컨트롤러만)

+++ src/main/java/com/nhnacademy/midoo/gateway/controller/task/MilestoneController.java
@@ -33,14 +33,12 @@ public class MilestoneController {
         taskService.putMilestone(milestoneId, milestonePutRequest);
 
         return "redirect:/myprojects";
-
     }
 
     @PostMapping("/{milestoneId}/delete")
-    public String deleteMilestone(@PathVariable int milestoneId) {
+    public String deleteMilestone(@PathVariable long milestoneId) {
         taskService.deleteMilestone(milestoneId);
 
         return "redirect:/myprojects";
-
     }
 }
